refactor: extract useMarkdownLinkHandler + preview builders (#507 step 8)
- useMarkdownLinkHandler: workspace-relative link click routing
- schedulerPreview / todoPreview: synthesize ToolResultComplete from
  raw JSON for the scheduler calendar / todo explorer views
- Drop relativeLink / scheduler / todo type-guard imports from FilesView

FilesView.vue final size: 245 lines (809 before #507).
