+++ src/components/FilesView.vue
@@ -58,17 +58,11 @@ import {
 } from "../composables/useFileSelection";
 import { useMarkdownMode } from "../composables/useMarkdownMode";
 import { useContentDisplay } from "../composables/useContentDisplay";
+import { useMarkdownLinkHandler } from "../composables/useMarkdownLinkHandler";
 import { apiPut } from "../utils/api";
 import { API_ROUTES } from "../config/apiRoutes";
-import { WORKSPACE_FILES } from "../config/workspacePaths";
-import type { ToolResultComplete } from "gui-chat-protocol/vue";
-import type { SchedulerData, ScheduledItem } from "../plugins/scheduler/index";
-import type { StatusColumn, TodoData, TodoItem } from "../plugins/todo/index";
-import {
-  isExternalHref,
-  resolveWorkspaceLink,
-  extractSessionIdFromPath,
-} from "../utils/path/relativeLink";
+import { toSchedulerResult } from "../utils/filesPreview/schedulerPreview";
+import { toTodoExplorerResult } from "../utils/filesPreview/todoPreview";
 
 const RECENT_THRESHOLD_MS = 60 * 1000;
 
@@ -164,80 +158,19 @@ watch(content, () => {
   rawSaveError.value = null;
 });
 
-function isScheduledItem(x: unknown): x is ScheduledItem {
-  if (typeof x !== "object" || x === null) return false;
-  if (!("id" in x) || typeof x.id !== "string") return false;
-  if (!("title" in x) || typeof x.title !== "string") return false;
-  return true;
-}
-
-function isScheduledItemArray(x: unknown): x is ScheduledItem[] {
-  return Array.isArray(x) && x.every(isScheduledItem);
-}
-
-// When the user opens the scheduler items file, render it with the
-// scheduler plugin's calendar view instead of as a JSON blob. We
-// synthesize a fake ToolResultComplete<SchedulerData> so the View
-// component receives the same shape it normally gets in chat mode.
-const schedulerResult = computed(
-  (): ToolResultComplete<SchedulerData> | null => {
-    if (selectedPath.value !== WORKSPACE_FILES.schedulerItems) return null;
-    if (!content.value || content.value.kind !== "text") return null;
-    let parsed: unknown;
-    try {
-      parsed = JSON.parse(content.value.content);
-    } catch {
-      return null;
-    }
-    if (!isScheduledItemArray(parsed)) return null;
-    return {
-      uuid: "files-scheduler-preview",
-      toolName: "manageScheduler",
-      message: WORKSPACE_FILES.schedulerItems,
-      title: "Scheduler",
-      data: { items: parsed },
-    };
-  },
+const schedulerResult = computed(() =>
+  toSchedulerResult(
+    selectedPath.value,
+    content.value?.kind === "text" ? content.value.content : null,
+  ),
 );
 
-// Same idea as schedulerResult: when the user opens the todos file
-// we render it as a full TodoExplorer (kanban / table / list) instead
-// of a raw JSON blob. The TodoExplorer fetches its own state from
-// /api/todos so the data we synthesize here is just a starter — the
-// columns array might be empty until the first refresh lands.
-function isTodoItem(x: unknown): x is TodoItem {
-  if (typeof x !== "object" || x === null) return false;
-  const o = x as Record<string, unknown>;
-  if (typeof o["id"] !== "string" || typeof o["text"] !== "string")
-    return false;
-  if (typeof o["completed"] !== "boolean") return false;
-  if (typeof o["createdAt"] !== "number") return false;
-  return true;
-}
-
-function isTodoItemArray(x: unknown): x is TodoItem[] {
-  return Array.isArray(x) && x.every(isTodoItem);
-}
-
-const todoExplorerResult = computed((): ToolResultComplete<TodoData> | null => {
-  if (selectedPath.value !== WORKSPACE_FILES.todosItems) return null;
-  if (!content.value || content.value.kind !== "text") return null;
-  let parsed: unknown;
-  try {
-    parsed = JSON.parse(content.value.content);
-  } catch {
-    return null;
-  }
-  const items: TodoItem[] = isTodoItemArray(parsed) ? parsed : [];
-  const columns: StatusColumn[] = [];
-  return {
-    uuid: "files-todo-preview",
-    toolName: "manageTodoList",
-    message: WORKSPACE_FILES.todosItems,
-    title: "Todo",
-    data: { items, columns },
-  };
-});
+const todoExplorerResult = computed(() =>
+  toTodoExplorerResult(
+    selectedPath.value,
+    content.value?.kind === "text" ? content.value.content : null,
+  ),
+);
 
 const recentPaths = computed(() => {
   const set = new Set<string>();
@@ -260,43 +193,12 @@ const recentPaths = computed(() => {
   return set;
 });
 
-// When the user clicks an <a> inside a rendered markdown body, check
-// if it's a workspace-internal relative/absolute link. If so, resolve
-// it against the current file and navigate inside FilesView instead
-// of letting the browser follow the (meaningless) relative href.
-//
 // Uses click.capture so we intercept before TextResponseView's own
 // handler (which only knows about absolute URLs) sees the event.
-function handleMarkdownLinkClick(event: MouseEvent): void {
-  if (event.button !== 0) return;
-  if (event.ctrlKey || event.metaKey || event.shiftKey) return;
-  const target = event.target as HTMLElement | null;
-  if (!target) return;
-  const anchor = target.closest("a");
-  if (!anchor) return;
-  const href = anchor.getAttribute("href");
-  if (!href) return;
-  // External URLs and mailto/tel: let TextResponseView's existing
-  // handler open them in a new tab.
-  if (isExternalHref(href)) return;
-  // Anchor-only (#section): let the browser handle in-page scroll.
-  if (href.startsWith("#")) return;
-  if (!selectedPath.value) return;
-  const resolved = resolveWorkspaceLink(selectedPath.value, href);
-  if (!resolved) return;
-  event.preventDefault();
-  event.stopPropagation();
-  // Chat session link: hand off to App.vue so the sidebar chat
-  // switches to that session instead of opening the raw jsonl
-  // as a file. Direct clicks in the file tree still open the
-  // jsonl in raw view — only markdown link clicks route here.
-  const sessionId = extractSessionIdFromPath(resolved);
-  if (sessionId !== null) {
-    emit("loadSession", sessionId);
-    return;
-  }
-  selectFile(resolved);
-}
+const { handleMarkdownLinkClick } = useMarkdownLinkHandler(selectedPath, {
+  onNavigate: selectFile,
+  onLoadSession: (sessionId) => emit("loadSession", sessionId),
+});
 
 // External URL changes (back/forward) → update selectedPath.
 watch(
